well..., got shit done, security is still todo

+++ src/main/java/hu/elte/alkfejl/alkfejl18/controllers/UserController.java
@@ -1,5 +1,9 @@
 package hu.elte.alkfejl.alkfejl18.controllers;
 
+import java.sql.Date;
+import java.sql.Time;
+import java.time.LocalTime;
+import java.util.ArrayList;
 import java.util.Optional;
 import hu.elte.alkfejl.alkfejl18.entities.*;
 import hu.elte.alkfejl.alkfejl18.repositories.*;
@@ -16,10 +20,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/user")
+@RequestMapping("/users")
 public class UserController {
     @Autowired
     private UserRepository userRepository;
+    @Autowired
+    private SkillRepository skillRepository;
     
     @GetMapping("")
     public ResponseEntity<Iterable<User>> getAll() {
@@ -28,9 +34,10 @@ public class UserController {
     }
      
     @PostMapping("/new")
-    public ResponseEntity<User> post(@RequestBody User user) {
-        Optional<User> oUser = userRepository.findByUsername(user.getUsername());
-        if (oUser.isPresent()) {
+    public ResponseEntity<User> createUser(@RequestBody User user) {
+        Optional<User> oUser = userRepository.findById(user.getId());
+        if (oUser.isPresent() || user.getOwnedProjects().size() > 0 || user.getProjects().size() > 0 ||
+        		user.getAssignedTasks().size() > 0) {
             return ResponseEntity.badRequest().build();
         }
         user.setId(null);
@@ -38,7 +45,7 @@ public class UserController {
     }
         
     @GetMapping("/{id}")
-    public ResponseEntity<User> get(@PathVariable Integer id) {
+    public ResponseEntity<User> getUser(@PathVariable Integer id) {
         Optional<User> oUser = userRepository.findById(id);
         if (!oUser.isPresent()) {
             return ResponseEntity.notFound().build();   
@@ -48,24 +55,87 @@ public class UserController {
     }
     
     @DeleteMapping("/{id}")
-    public ResponseEntity delete(@PathVariable Integer id) {
+    public ResponseEntity deleteUser(@PathVariable Integer id) {
         Optional<User> oUser = userRepository.findById(id);
         if (!oUser.isPresent()) {
             return ResponseEntity.notFound().build();   
         }
-            
+        if(oUser.get().getOwnedProjects().size() > 0 || oUser.get().getProjects().size() > 0 ||
+        		oUser.get().getAssignedTasks().size() > 0) {
+        	return ResponseEntity.badRequest().build();
+        }
         userRepository.delete(oUser.get());
         return ResponseEntity.ok().build();
     }
-     
-    @PutMapping("/edit/{id}")
-    public ResponseEntity<User> put(@PathVariable Integer id, @RequestBody User user) {
-        Optional<User> oUser = userRepository.findById(id);
+    
+    @PutMapping("/{id}/edit")
+    public ResponseEntity<User> editUser(@PathVariable Integer id,@RequestBody User user){
+    	if(user.getAssignedTasks() != null || user.getSkills() != null ||
+    				user.getOwnedProjects() != null || user.getProjects() != null) {
+    		return ResponseEntity.badRequest().build();
+    	}
+    	Optional<User> oUser = userRepository.findById(id);
         if (!oUser.isPresent()) {
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.notFound().build();   
         }
         
         user.setId(id);
         return ResponseEntity.ok(userRepository.save(user));
     }
+    
+    @GetMapping("/{id}/skills")
+    public ResponseEntity<Iterable<Skill>> getSkillList(@PathVariable Integer id){
+    	 Optional<User> oUser = userRepository.findById(id);
+         if (!oUser.isPresent()) {
+             return ResponseEntity.notFound().build();
+         }
+
+         return ResponseEntity.ok(oUser.get().getSkills());
+    }
+    
+    @PutMapping("/{id}/skills")
+    public ResponseEntity<User> addSkill(@PathVariable Integer id, @RequestBody Skill skill) {
+        Optional<User> oUser = userRepository.findById(id);
+        Optional<Skill> oSkill = skillRepository.findByName(skill.getName());
+        if (!oUser.isPresent() || !oSkill.isPresent()) {
+            return ResponseEntity.notFound().build();
+        }
+        oUser.get().getSkills().add(skill);
+        skill.getOwners().add(oUser.get());
+        skillRepository.save(skill);
+        return ResponseEntity.ok(userRepository.save(oUser.get()));
+    }
+    
+    @GetMapping("/{id}/ownedProjects")
+    public ResponseEntity<Iterable<Project>> getOwnedProjectList(@PathVariable Integer id){
+    	 Optional<User> oUser = userRepository.findById(id);
+         if (!oUser.isPresent()) {
+             return ResponseEntity.notFound().build();
+         }
+
+         return ResponseEntity.ok(oUser.get().getOwnedProjects());
+    }
+    
+    @GetMapping("/{id}/projects")
+    public ResponseEntity<Iterable<Project>> getProjectList(@PathVariable Integer id){
+    	 Optional<User> oUser = userRepository.findById(id);
+         if (!oUser.isPresent()) {
+             return ResponseEntity.notFound().build();
+         }
+
+         return ResponseEntity.ok(oUser.get().getProjects());
+    }
+    
+    @GetMapping("/{id}/assignedTasks")
+    public ResponseEntity<Iterable<Task>> getTaskList(@PathVariable Integer id){
+    	 Optional<User> oUser = userRepository.findById(id);
+         if (!oUser.isPresent()) {
+             return ResponseEntity.notFound().build();
+         }
+
+         return ResponseEntity.ok(oUser.get().getAssignedTasks());
+    }
+    
+    
+    
 }
